refactor(auth): rename Node runtime subpath
