+++ .changeset/rename-auth-node-runtime.md
@@ -1,0 +1,5 @@
+---
+"@voyant-travel/auth": minor
+---
+
+Rename the Node runtime subpath from `@voyant-travel/auth/operator-node-runtime` to `@voyant-travel/auth/node-runtime`.
